requirements.txt
# AI Email Agent - Streamlined Requirements
# Python 3.10+ required

# ==================== CORE WEB FRAMEWORK ====================
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
python-multipart>=0.0.6

# ==================== GOOGLE API INTEGRATIONS ====================
google-api-python-client>=2.180.0
google-auth>=2.23.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-generativeai>=0.3.2

# ==================== DATA MODELS & VALIDATION ====================
pydantic>=2.11.0
pydantic-settings>=2.10.0
python-dotenv>=1.0.0

# ==================== MACHINE LEARNING - YOUR TRAINED MODELS ====================
# PyTorch (CUDA support for your models)
torch>=2.8.0

# HuggingFace Transformers (for YOUR DistilBERT classifier)
transformers>=4.56.0

# Model optimization (for quantized models)
accelerate>=1.10.0
bitsandbytes>=0.47.0

# LoRA fine-tuning (for YOUR Llama 3 extractor)
peft>=0.17.0

# Text processing (for Llama tokenizer)
sentencepiece>=0.2.0

# ==================== HTTP & NETWORKING ====================
requests>=2.32.0
aiohttp>=3.12.0

# ==================== DATE/TIME HANDLING ====================
python-dateutil>=2.9.0

# ==================== UTILITIES ====================
tqdm>=4.67.0
tenacity>=9.1.0
